# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Geenbank Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dossiers/me/intake:
    put:
      operationId: updateMyIntake
      tags: [dossiers]
      summary: Update intake fields on the prospect's own dossier
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IntakeUpdate" }
      responses:
        "200":
          description: Updated dossier
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dossier" }
...
  /dossiers:
    get:
      operationId: listDossiers
      tags: [dossiers]
      summary: List dossiers (loan officer / admin)
      parameters:
        - in: query
          name: bucket
          required: false
          schema:
            type: string
            enum: [all, new, in_review, ready, additional_info, approved, rejected]
      responses:
        "200":
          description: Dossiers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DossierListItem" }
  /dossiers/{dossierId}:
    get:
      operationId: getDossier
      tags: [dossiers]
      summary: Get a dossier by id (loan officer / admin)
      parameters:
        - in: path
          name: dossierId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Dossier
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dossier" }
  /dossiers/{dossierId}/decision:
    post:
      operationId: makeDossierDecision
      tags: [dossiers]
      summary: Approve, reject, or request additional information
      parameters:
        - in: path
          name: dossierId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DecisionRequest" }
      responses:
        "200":
          description: Updated dossier
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dossier" }
  /dossiers/me/documents:
    get:
      operationId: listMyDocuments
      tags: [documents]
      summary: List the prospect's own documents
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Document" }
    post:
      operationId: uploadMyDocument
      tags: [documents]
      summary: Upload a document to the prospect's own dossier (mock storage)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DocumentUpload" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
  /dossiers/me/documents/{documentId}:
    delete:
      operationId: deleteMyDocument
      tags: [documents]
      summary: Delete a document from the prospect's own dossier
      parameters:
        - in: path
          name: documentId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /dossiers/{dossierId}/documents:
    get:
      operationId: listDossierDocuments
      tags: [documents]
      summary: List documents on a specific dossier (loan officer / admin)
      parameters:
        - in: path
          name: dossierId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Document" }
...
  /partners:
    get:
      operationId: listPartners
      tags: [partners]
      summary: List partner financiers
      responses:
        "200":
          description: Partners
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PartnerFinancier" }
    post:
      operationId: createPartner
      tags: [partners]
      summary: Create a partner financier
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PartnerInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PartnerFinancier" }
  /partners/{partnerId}:
    put:
      operationId: updatePartner
      tags: [partners]
      summary: Update a partner financier
      parameters:
        - in: path
          name: partnerId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PartnerInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PartnerFinancier" }
...
components:
  schemas:
...
    Ok:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
...
    DossierStatus:
      type: string
      enum:
        - lead_created
        - account_invited
        - prospect_logged_in
        - intake_in_progress
        - documents_uploaded
        - pre_validation_running
        - blocked_missing_documents
        - blocked_invalid_documents
        - ready_for_ai_analysis
        - ai_analysis_running
        - entrepreneur_report_ready
        - submitted_to_geenbank
        - loan_officer_review
        - additional_info_requested
        - approved_for_partner_submission
        - rejected_by_loan_officer
        - memorandum_generated
        - submitted_to_partners
        - partner_response_received
        - closed
...
    ProspectProfile:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        companyName: { type: string }
        contactName: { type: string }
        kvkNumber: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        source: { type: string, nullable: true }
        pipedriveDealId: { type: string, nullable: true }
        createdAt: { type: string }
      required: [id, userId, companyName, contactName, createdAt]
    Dossier:
      type: object
      properties:
        id: { type: string }
        prospectId: { type: string }
        prospect: { $ref: "#/components/schemas/ProspectProfile" }
        status: { $ref: "#/components/schemas/DossierStatus" }
        currentStage: { type: string }
        financingPurpose: { type: string, nullable: true }
        requestedAmount: { type: number, nullable: true }
        financingTypePreference: { type: string, nullable: true }
        existingFinancing: { type: string, nullable: true }
        annualRevenue: { type: number, nullable: true }
        annualCost: { type: number, nullable: true }
        annualProfit: { type: number, nullable: true }
        companyDescription: { type: string, nullable: true }
        completenessScore: { type: number, nullable: true }
        correctnessScore: { type: number, nullable: true }
        viabilityScore: { type: number, nullable: true }
        confidenceScore: { type: number, nullable: true }
        aiVerdict: { type: string, nullable: true }
        loanOfficerDecision: { type: string, nullable: true }
        loanOfficerNotes: { type: string, nullable: true }
        submittedAt: { type: string, nullable: true }
        createdAt: { type: string }
        updatedAt: { type: string }
        intakeCompletionPercent: { type: number }
        documentsCount: { type: integer }
        blockingConditionsCount: { type: integer }
      required:
        - id
        - prospectId
        - status
        - currentStage
        - createdAt
        - updatedAt
        - intakeCompletionPercent
        - documentsCount
        - blockingConditionsCount
    DossierListItem:
      type: object
      properties:
        id: { type: string }
        companyName: { type: string }
        contactName: { type: string }
        status: { $ref: "#/components/schemas/DossierStatus" }
        currentStage: { type: string }
        requestedAmount: { type: number, nullable: true }
        financingPurpose: { type: string, nullable: true }
        completenessScore: { type: number, nullable: true }
        correctnessScore: { type: number, nullable: true }
        viabilityScore: { type: number, nullable: true }
        confidenceScore: { type: number, nullable: true }
        aiVerdict: { type: string, nullable: true }
        submittedAt: { type: string, nullable: true }
        updatedAt: { type: string }
        bucket: { type: string }
      required:
        - id
        - companyName
        - contactName
        - status
        - currentStage
        - updatedAt
        - bucket
    IntakeUpdate:
      type: object
      properties:
        companyName: { type: string, nullable: true }
        contactName: { type: string, nullable: true }
        kvkNumber: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        financingPurpose: { type: string, nullable: true }
        requestedAmount: { type: number, nullable: true }
        financingTypePreference: { type: string, nullable: true }
        existingFinancing: { type: string, nullable: true }
        annualRevenue: { type: number, nullable: true }
        annualCost: { type: number, nullable: true }
        annualProfit: { type: number, nullable: true }
        companyDescription: { type: string, nullable: true }
    Document:
      type: object
      properties:
        id: { type: string }
        dossierId: { type: string }
        uploadedBy: { type: string }
        documentType: { type: string }
        filename: { type: string }
        mimeType: { type: string }
        sizeBytes: { type: integer }
        storagePath: { type: string }
        uploadStatus: { type: string }
        validationStatus: { type: string }
        extractedDataStatus: { type: string }
        usedInAnalysis: { type: boolean }
        validationNotes: { type: string, nullable: true }
        createdAt: { type: string }
      required:
        - id
        - dossierId
        - uploadedBy
        - documentType
        - filename
        - mimeType
        - sizeBytes
        - uploadStatus
        - validationStatus
        - extractedDataStatus
        - usedInAnalysis
        - createdAt
    DocumentUpload:
      type: object
      properties:
        documentType: { type: string }
        filename: { type: string }
        mimeType: { type: string }
        sizeBytes: { type: integer }
        contentBase64: { type: string, nullable: true }
      required: [documentType, filename, mimeType, sizeBytes]
...
    DecisionRequest:
      type: object
      properties:
        decision:
          type: string
          enum: [approve, reject, request_additional_info]
        notes: { type: string, nullable: true }
        requestedItems:
          type: array
          items: { type: string }
      required: [decision]
    PartnerFinancier:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        contactEmail: { type: string }
        productFocus: { type: string }
        minimumTicketSize: { type: number, nullable: true }
        maximumTicketSize: { type: number, nullable: true }
        activeStatus: { type: string }
        notes: { type: string, nullable: true }
        submissionsCount: { type: integer }
        acceptedCount: { type: integer }
        createdAt: { type: string }
      required:
        - id
        - name
        - contactEmail
        - productFocus
        - activeStatus
        - submissionsCount
        - acceptedCount
        - createdAt
    PartnerInput:
      type: object
      properties:
        name: { type: string }
        contactEmail: { type: string }
        productFocus: { type: string }
        minimumTicketSize: { type: number, nullable: true }
        maximumTicketSize: { type: number, nullable: true }
        activeStatus: { type: string }
        notes: { type: string, nullable: true }
      required: [name, contactEmail, productFocus, activeStatus]
...